Test: moved inital indexedDB set up in index page. Trying to get indexedDB to work on Safari
